Implement dummy for other methods, reorganized files so that host functions are in the proper file
