+++ internal/array.go
@@ -3,6 +3,7 @@ package embind
 import (
 	"context"
 	"fmt"
+
 	"github.com/tetratelabs/wazero/api"
 )
 
@@ -87,3 +88,134 @@ func (at *arrayType) DestructorFunction(ctx context.Context, mod api.Module, poi
 func (at *arrayType) GoType() string {
 	return "[]any"
 }
+
+var RegisterValueArray = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	rawType := api.DecodeI32(stack[0])
+	namePtr := api.DecodeI32(stack[1])
+	constructorSignature := api.DecodeI32(stack[2])
+	rawConstructor := api.DecodeI32(stack[3])
+	destructorSignature := api.DecodeI32(stack[4])
+	rawDestructor := api.DecodeI32(stack[5])
+
+	name, err := engine.readCString(uint32(namePtr))
+	if err != nil {
+		panic(fmt.Errorf("could not read name: %w", err))
+	}
+
+	rawConstructorFunc, err := engine.newInvokeFunc(constructorSignature, rawConstructor, []api.ValueType{}, []api.ValueType{api.ValueTypeI32})
+	if err != nil {
+		panic(fmt.Errorf("could not create rawConstructorFunc: %w", err))
+	}
+
+	rawDestructorFunc, err := engine.newInvokeFunc(destructorSignature, rawDestructor, []api.ValueType{api.ValueTypeI32}, []api.ValueType{})
+	if err != nil {
+		panic(fmt.Errorf("could not create rawDestructorFunc: %w", err))
+	}
+
+	engine.registeredTuples[rawType] = &registeredTuple{
+		name:           name,
+		rawConstructor: rawConstructorFunc,
+		rawDestructor:  rawDestructorFunc,
+		elements:       []*registeredTupleElement{},
+	}
+})
+
+var RegisterValueArrayElement = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	rawTupleType := api.DecodeI32(stack[0])
+	getterReturnType := api.DecodeI32(stack[1])
+	getterSignature := api.DecodeI32(stack[2])
+	getter := api.DecodeI32(stack[3])
+	getterContext := api.DecodeI32(stack[4])
+	setterArgumentType := api.DecodeI32(stack[5])
+	setterSignature := api.DecodeI32(stack[6])
+	setter := api.DecodeI32(stack[7])
+	setterContext := api.DecodeI32(stack[8])
+
+	engine.registeredTuples[rawTupleType].elements = append(engine.registeredTuples[rawTupleType].elements, &registeredTupleElement{
+		getter:             getter,
+		getterSignature:    getterSignature,
+		getterReturnType:   getterReturnType,
+		getterContext:      getterContext,
+		setter:             setter,
+		setterSignature:    setterSignature,
+		setterArgumentType: setterArgumentType,
+		setterContext:      setterContext,
+	})
+})
+
+var FinalizeValueArray = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	rawTupleType := api.DecodeI32(stack[0])
+	reg := engine.registeredTuples[rawTupleType]
+	delete(engine.registeredTuples, rawTupleType)
+	elements := reg.elements
+	elementsLength := len(elements)
+
+	elementTypes := make([]int32, len(elements)*2)
+	for i := range elements {
+		elementTypes[i] = elements[i].getterReturnType
+		elementTypes[i+len(elements)] = elements[i].setterArgumentType
+	}
+
+	err := engine.whenDependentTypesAreResolved([]int32{rawTupleType}, elementTypes, func(types []registeredType) ([]registeredType, error) {
+		for i := range elements {
+			element := elements[i]
+			getterReturnType := types[i]
+
+			getterFunc, err := engine.newInvokeFunc(element.getterSignature, element.getter, []api.ValueType{api.ValueTypeI32, api.ValueTypeI32}, []api.ValueType{getterReturnType.NativeType()})
+			if err != nil {
+				return nil, fmt.Errorf("could not create getterFunc: %w", err)
+			}
+
+			setterArgumentType := types[i+len(elements)]
+			setterFunc, err := engine.newInvokeFunc(element.setterSignature, element.setter, []api.ValueType{api.ValueTypeI32, api.ValueTypeI32, setterArgumentType.NativeType()}, []api.ValueType{})
+			if err != nil {
+				return nil, fmt.Errorf("could not create setterFunc: %w", err)
+			}
+
+			element.read = func(ctx context.Context, mod api.Module, ptr int32) (any, error) {
+				res, err := getterFunc.Call(ctx, api.EncodeI32(element.getterContext), api.EncodeI32(ptr))
+				if err != nil {
+					return nil, err
+				}
+				return getterReturnType.FromWireType(ctx, mod, res[0])
+			}
+			element.write = func(ctx context.Context, mod api.Module, ptr int32, o any) error {
+				destructors := &[]*destructorFunc{}
+				res, err := setterArgumentType.ToWireType(ctx, mod, destructors, o)
+				if err != nil {
+					return err
+				}
+
+				_, err = setterFunc.Call(ctx, api.EncodeI32(element.setterContext), api.EncodeI32(ptr), res)
+				if err != nil {
+					return err
+				}
+
+				err = engine.runDestructors(ctx, *destructors)
+				if err != nil {
+					return err
+				}
+
+				return nil
+			}
+		}
+
+		return []registeredType{
+			&arrayType{
+				baseType: baseType{
+					rawType:        rawTupleType,
+					name:           reg.name,
+					argPackAdvance: 8,
+				},
+				reg:            reg,
+				elementsLength: elementsLength,
+			},
+		}, nil
+	})
+	if err != nil {
+		panic(fmt.Errorf("could not call whenDependentTypesAreResolved: %w", err))
+	}
+})
