+++ internal/constant.go
@@ -1,7 +1,10 @@
 package embind
 
 import (
+	"context"
 	"fmt"
+
+	"github.com/tetratelabs/wazero/api"
 )
 
 type IConstant interface {
@@ -49,3 +52,47 @@ func (e *engine) GetConstants() []IConstant {
 	}
 	return constants
 }
+
+var RegisterConstant = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	name, err := engine.readCString(uint32(api.DecodeI32(stack[0])))
+	if err != nil {
+		panic(fmt.Errorf("could not read name: %w", err))
+	}
+
+	rawType := api.DecodeI32(stack[1])
+	constantValue := api.DecodeF64(stack[2])
+
+	err = engine.whenDependentTypesAreResolved([]int32{}, []int32{rawType}, func(argTypes []registeredType) ([]registeredType, error) {
+		registeredType := argTypes[0]
+		// We need to do this since the JS VM automatically converts between
+		// the float64 and other types, but we can't do this, we need to
+		// manually convert everything. Note that the value inside the original
+		// uint64 is an actual typecast F64, so it's not that the stack
+		// contains FromWireType'able data for the given type.
+		cppValue := registeredType.FromF64(constantValue)
+		val, err := registeredType.FromWireType(ctx, engine.mod, cppValue)
+		if err != nil {
+			return nil, fmt.Errorf("could not initialize constant %s: %w", name, err)
+		}
+
+		_, ok := engine.registeredConstants[name]
+		if !ok {
+			engine.registeredConstants[name] = &registeredConstant{
+				name: name,
+			}
+		}
+
+		engine.registeredConstants[name].registeredType = registeredType
+		engine.registeredConstants[name].hasCppValue = true
+		engine.registeredConstants[name].cppValue = val
+		engine.registeredConstants[name].rawCppValue = cppValue
+
+		return nil, engine.registeredConstants[name].validate()
+	})
+
+	if err != nil {
+		panic(err)
+	}
+})
