+++ internal/embind.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+
 	"github.com/tetratelabs/wazero/api"
 )
