+++ internal/emval.go
@@ -2,8 +2,11 @@ package embind
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"reflect"
+	"strconv"
+	"strings"
 	"unicode"
 
 	"github.com/jerbob92/wazero-emscripten-embind/types"
@@ -408,3 +411,569 @@ func (e *emvalEngine) callMethod(ctx context.Context, mod api.Module, registered
 
 	return res, nil
 }
+
+var RegisterEmval = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	rawType := api.DecodeI32(stack[0])
+	name, err := engine.readCString(uint32(api.DecodeI32(stack[1])))
+	if err != nil {
+		panic(fmt.Errorf("could not read name: %w", err))
+	}
+
+	err = engine.registerType(rawType, &emvalType{
+		baseType: baseType{
+			rawType:        rawType,
+			name:           name,
+			argPackAdvance: 8,
+		},
+	}, &registerTypeOptions{
+		ignoreDuplicateRegistrations: true,
+	})
+	if err != nil {
+		panic(fmt.Errorf("could not register: %w", err))
+	}
+})
+
+var EmvalTakeValue = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	rawType := api.DecodeI32(stack[0])
+
+	registeredType, ok := engine.registeredTypes[rawType]
+	if !ok {
+		typeName, err := engine.getTypeName(ctx, rawType)
+		if err != nil {
+			panic(err)
+		}
+		panic(fmt.Errorf("_emval_take_value has unknown type %s", typeName))
+	}
+
+	arg := api.DecodeI32(stack[1])
+	value, err := registeredType.ReadValueFromPointer(ctx, mod, uint32(arg))
+	if err != nil {
+		panic(fmt.Errorf("could not take value for _emval_take_value: %w", err))
+	}
+
+	id := engine.emvalEngine.toHandle(value)
+	stack[0] = api.EncodeI32(id)
+})
+
+var EmvalIncref = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	handle := api.DecodeI32(stack[0])
+	err := engine.emvalEngine.allocator.incref(handle)
+	if err != nil {
+		panic(fmt.Errorf("could not emval incref: %w", err))
+	}
+})
+
+var EmvalDecref = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	handle := api.DecodeI32(stack[0])
+	err := engine.emvalEngine.allocator.decref(handle)
+	if err != nil {
+		panic(fmt.Errorf("could not emval incref: %w", err))
+	}
+})
+
+var EmvalRegisterSymbol = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	address := uint32(api.DecodeI32(stack[0]))
+	name, err := engine.readCString(address)
+	if err != nil {
+		panic(fmt.Errorf("could not get symbol name"))
+	}
+	engine.emvalEngine.symbols[address] = name
+})
+
+var EmvalGetGlobal = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	name := api.DecodeI32(stack[0])
+
+	if name == 0 {
+		stack[0] = api.EncodeI32(engine.emvalEngine.toHandle(engine.emvalEngine.getGlobal("")))
+	} else {
+		name, err := engine.getStringOrSymbol(uint32(name))
+		if err != nil {
+			panic(fmt.Errorf("could not get symbol name"))
+		}
+		stack[0] = api.EncodeI32(engine.emvalEngine.toHandle(engine.emvalEngine.getGlobal(name)))
+	}
+})
+
+var EmvalAs = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	id := api.DecodeI32(stack[0])
+	handle, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	returnType, err := engine.requireRegisteredType(ctx, api.DecodeI32(stack[1]), "emval::as")
+	if err != nil {
+		panic(fmt.Errorf("could not require registered type: %w", err))
+	}
+
+	var destructors = &[]*destructorFunc{}
+	rd := engine.emvalEngine.toHandle(destructors)
+	ok := mod.Memory().WriteUint32Le(uint32(api.DecodeI32(stack[2])), uint32(rd))
+	if !ok {
+		panic(fmt.Errorf("could not write destructor ref to memory"))
+	}
+
+	returnVal, err := returnType.ToWireType(ctx, mod, destructors, handle)
+	if err != nil {
+		panic(fmt.Errorf("could not call toWireType on _emval_as: %w", err))
+	}
+
+	stack[0] = returnVal
+})
+
+var EmvalNew = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	id := api.DecodeI32(stack[0])
+
+	handle, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not get value of handle: %w", err))
+	}
+
+	argCount := int(api.DecodeI32(stack[1]))
+	argsTypeBase := uint32(api.DecodeI32(stack[2]))
+	argsBase := uint32(api.DecodeI32(stack[3]))
+
+	args := make([]any, argCount)
+	argTypeNames := make([]string, argCount)
+	for i := 0; i < argCount; i++ {
+		argType, ok := mod.Memory().ReadUint32Le(argsTypeBase + (4 * uint32(i)))
+		if !ok {
+			panic(fmt.Errorf("could not read arg type for arg %d from memory", i))
+		}
+
+		registeredArgType, err := engine.requireRegisteredType(ctx, int32(argType), fmt.Sprintf("argument %d", i))
+		if err != nil {
+			panic(fmt.Errorf("could not require registered type: %w", err))
+		}
+
+		args[i], err = registeredArgType.ReadValueFromPointer(ctx, mod, argsBase+(8*uint32(i)))
+		if err != nil {
+			panic(fmt.Errorf("could not read arg value from pointer for arg %d, %w", i, err))
+		}
+
+		argTypeNames[i] = registeredArgType.Name()
+	}
+
+	var res any
+	c, ok := handle.(IEmvalConstructor)
+	if ok {
+		res, err = c.New(argTypeNames, args...)
+		if err != nil {
+			panic(fmt.Errorf("could not instaniate new value on %T with New(): %w", handle, err))
+		}
+	} else {
+		typeElem := reflect.TypeOf(handle)
+
+		// If we received a pointer, resolve the struct behind it.
+		if typeElem.Kind() == reflect.Pointer {
+			typeElem = typeElem.Elem()
+		}
+
+		// Make new instance of struct.
+		newElem := reflect.New(typeElem)
+
+		// Set the values on the struct if we need to/can.
+		if argCount > 0 {
+			if typeElem.Kind() != reflect.Struct {
+				panic(fmt.Errorf("could not instaniate new value of %T: arguments required but can only be set on a struct", handle))
+			}
+
+			for i := 0; i < argCount; i++ {
+				for fieldI := 0; fieldI < typeElem.NumField(); fieldI++ {
+					var err error
+					func() {
+						defer func() {
+							if recoverErr := recover(); recoverErr != nil {
+								realError, ok := recoverErr.(error)
+								if ok {
+									err = fmt.Errorf("could not set arg %d with embind_arg tag on emval %T: %w", i, handle, realError)
+								}
+								err = fmt.Errorf("could not set arg %d with embind_arg tag on emval %T: %v", i, handle, recoverErr)
+							}
+						}()
+
+						val := typeElem.Field(fieldI)
+						if val.Tag.Get("embind_arg") == strconv.Itoa(i) {
+							f := newElem.Elem().FieldByName(val.Name)
+							if f.IsValid() && f.CanSet() {
+								f.Set(reflect.ValueOf(args[i]))
+							}
+						}
+					}()
+					if err != nil {
+						panic(fmt.Errorf("could not instaniate new value of %T: %w", handle, err))
+					}
+				}
+			}
+		}
+
+		if reflect.TypeOf(handle).Kind() == reflect.Pointer {
+			res = newElem.Interface()
+		} else {
+			res = newElem.Elem().Interface()
+		}
+	}
+
+	stack[0] = api.EncodeI32(engine.emvalEngine.toHandle(res))
+})
+
+var EmvalSetProperty = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	handle, err := engine.emvalEngine.toValue(api.DecodeI32(stack[0]))
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	key, err := engine.emvalEngine.toValue(api.DecodeI32(stack[1]))
+	if err != nil {
+		panic(fmt.Errorf("could not find key: %w", err))
+	}
+
+	val, err := engine.emvalEngine.toValue(api.DecodeI32(stack[2]))
+	if err != nil {
+		panic(fmt.Errorf("could not find val: %w", err))
+	}
+
+	keyString, ok := key.(string)
+	if !ok {
+		panic(fmt.Errorf("could not set property on emval %T: %w", handle, errors.New("key is not of type string")))
+	}
+
+	f, err := engine.emvalEngine.getElemField(handle, keyString)
+	if err != nil {
+		panic(fmt.Errorf("could not set property %s on emval %T: %w", keyString, handle, err))
+	}
+
+	defer func() {
+		if err := recover(); err != nil {
+			realError, ok := err.(error)
+			if ok {
+				panic(fmt.Errorf("could not set property %s on emval %T: %w", keyString, handle, realError))
+			}
+			panic(fmt.Errorf("could not set property %s on emval %T: %v", keyString, handle, err))
+		}
+	}()
+
+	f.Set(reflect.ValueOf(val))
+})
+
+var EmvalGetProperty = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	handle, err := engine.emvalEngine.toValue(api.DecodeI32(stack[0]))
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	key, err := engine.emvalEngine.toValue(api.DecodeI32(stack[1]))
+	if err != nil {
+		panic(fmt.Errorf("could not find key: %w", err))
+	}
+
+	keyString, ok := key.(string)
+	if !ok {
+		panic(fmt.Errorf("could not get property on emval %T: %w", handle, errors.New("key is not of type string")))
+	}
+
+	f, err := engine.emvalEngine.getElemField(handle, keyString)
+	if err != nil {
+		panic(fmt.Errorf("could not get property %s on emval %T: %w", keyString, handle, err))
+	}
+
+	defer func() {
+		if err := recover(); err != nil {
+			realError, ok := err.(error)
+			if ok {
+				panic(fmt.Errorf("could not get property %s on emval %T: %w", keyString, handle, realError))
+			}
+			panic(fmt.Errorf("could not get property %s on emval %T: %v", keyString, handle, err))
+		}
+	}()
+
+	stack[0] = api.EncodeI32(engine.emvalEngine.toHandle(f.Interface()))
+})
+
+var EmvalNewCString = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	v := api.DecodeI32(stack[0])
+	name, err := engine.getStringOrSymbol(uint32(v))
+	if err != nil {
+		panic(fmt.Errorf("could not get symbol name"))
+	}
+	stack[0] = api.EncodeI32(engine.emvalEngine.toHandle(name))
+})
+
+var EmvalRunDestructors = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	id := api.DecodeI32(stack[0])
+	destructorsVal, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	destructors := destructorsVal.(*[]*destructorFunc)
+
+	err = engine.runDestructors(ctx, *destructors)
+	if err != nil {
+		panic(fmt.Errorf("could not run destructors: %w", err))
+	}
+
+	err = engine.emvalEngine.allocator.decref(id)
+	if err != nil {
+		panic(fmt.Errorf("could not run decref id %d: %w", id, err))
+	}
+})
+
+var EmvalGetMethodCaller = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	argCount := int(api.DecodeI32(stack[0]))
+	argsTypeBase := uint32(api.DecodeI32(stack[1]))
+
+	typeNames := make([]string, argCount)
+	argTypes := make([]registeredType, argCount)
+	for i := 0; i < argCount; i++ {
+		argType, ok := mod.Memory().ReadUint32Le(argsTypeBase + (4 * uint32(i)))
+		if !ok {
+			panic(fmt.Errorf("could not read arg type for arg %d from memory", i))
+		}
+
+		registeredType, err := engine.requireRegisteredType(ctx, int32(argType), fmt.Sprintf("argument %d", i))
+		if err != nil {
+			panic(fmt.Errorf("could not require registered type: %w", err))
+		}
+
+		typeNames[i] = registeredType.Name()
+		argTypes[i] = registeredType
+	}
+
+	signatureName := typeNames[0] + "_$" + strings.Join(typeNames[1:], "_") + "$"
+
+	id, ok := engine.emvalEngine.registeredMethodIds[signatureName]
+	if ok {
+		stack[0] = api.EncodeI32(id)
+		return
+	}
+
+	newID := engine.emvalEngine.registeredMethodCount
+	newRegisteredMethod := &emvalRegisteredMethod{
+		id:       newID,
+		argTypes: argTypes,
+		name:     signatureName,
+	}
+	engine.emvalEngine.registeredMethodIds[signatureName] = newID
+	engine.emvalEngine.registeredMethods[newID] = newRegisteredMethod
+	engine.emvalEngine.registeredMethodCount++
+
+	stack[0] = api.EncodeI32(newID)
+	return
+})
+
+var EmvalCall = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	id := api.DecodeI32(stack[0])
+	argCount := api.DecodeI32(stack[1])
+	argTypes := api.DecodeI32(stack[2])
+	argv := api.DecodeI32(stack[3])
+
+	handle, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	types, err := engine.lookupTypes(ctx, argCount, argTypes)
+	if err != nil {
+		panic(fmt.Errorf("could not load required types: %w", err))
+	}
+
+	args := make([]any, argCount)
+	for i := 0; i < int(argCount); i++ {
+		requiredType := types[i]
+		args[i], err = requiredType.ReadValueFromPointer(ctx, mod, uint32(argv))
+		if err != nil {
+			panic(fmt.Errorf("could not load argument value: %w", err))
+		}
+
+		argv += requiredType.ArgPackAdvance()
+	}
+
+	// @todo: implement me.
+	//rv := handle.apply(undefined, args)
+	rv := struct{}{}
+	newHandle := engine.emvalEngine.toHandle(rv)
+	stack[0] = api.EncodeI32(newHandle)
+	panic(fmt.Errorf("Emval call unimplemented %v", handle))
+})
+
+var EmvalCallMethod = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	caller := api.DecodeI32(stack[0])
+
+	registeredMethod, ok := engine.emvalEngine.registeredMethods[caller]
+	if !ok {
+		panic(fmt.Errorf("could not call method with ID %d", caller))
+	}
+
+	id := api.DecodeI32(stack[1])
+	handle, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	methodName, err := engine.getStringOrSymbol(uint32(api.DecodeI32(stack[2])))
+	if err != nil {
+		panic(fmt.Errorf("could not get symbol name"))
+	}
+
+	argsBase := uint32(api.DecodeI32(stack[4]))
+	destructorsRef := uint32(api.DecodeI32(stack[3]))
+
+	res, err := engine.emvalEngine.callMethod(ctx, mod, registeredMethod, handle, methodName, destructorsRef, argsBase)
+	if err != nil {
+		panic(fmt.Errorf("could not call %s on %T: %w", methodName, handle, err))
+	}
+	stack[0] = api.EncodeF64(float64(res))
+})
+
+var EmvalCallVoidMethod = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+	caller := api.DecodeI32(stack[0])
+
+	registeredMethod, ok := engine.emvalEngine.registeredMethods[caller]
+	if !ok {
+		panic(fmt.Errorf("could not call method with ID %d", caller))
+	}
+
+	id := api.DecodeI32(stack[1])
+	handle, err := engine.emvalEngine.toValue(id)
+	if err != nil {
+		panic(fmt.Errorf("could not find handle: %w", err))
+	}
+
+	methodName, err := engine.getStringOrSymbol(uint32(api.DecodeI32(stack[2])))
+	if err != nil {
+		panic(fmt.Errorf("could not get symbol name"))
+	}
+
+	argsBase := uint32(api.DecodeI32(stack[3]))
+
+	_, err = engine.emvalEngine.callMethod(ctx, mod, registeredMethod, handle, methodName, 0, argsBase)
+	if err != nil {
+		panic(fmt.Errorf("could not call %s on %T: %w", methodName, handle, err))
+	}
+})
+
+var EmvalInstanceof = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalInstanceof call unimplemented")
+})
+
+var EmvalTypeof = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalTypeof call unimplemented")
+})
+
+var EmvalAsInt64 = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalAsInt64 call unimplemented")
+})
+
+var EmvalAsUint64 = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalAsUint64 call unimplemented")
+})
+
+var EmvalAwait = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalAwait call unimplemented")
+})
+
+var EmvalDelete = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalDelete call unimplemented")
+})
+
+var EmvalEquals = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalEquals call unimplemented")
+})
+
+var EmvalGetModuleProperty = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalGetModuleProperty call unimplemented")
+})
+
+var EmvalIn = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalIn call unimplemented")
+})
+
+var EmvalIsNumber = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalIsNumber call unimplemented")
+})
+
+var EmvalIsString = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalIsString call unimplemented")
+})
+
+var EmvalLessThan = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalLessThan call unimplemented")
+})
+
+var EmvalNewArray = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNewArray call unimplemented")
+})
+
+var EmvalNewArrayFromMemoryView = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNewArrayFromMemoryView call unimplemented")
+})
+
+var EmvalNewObject = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNewObject call unimplemented")
+})
+
+var EmvalNewU16string = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNewU16string call unimplemented")
+})
+
+var EmvalNewU8string = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNewU8string call unimplemented")
+})
+
+var EmvalNot = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalNot call unimplemented")
+})
+
+var EmvalStrictlyEquals = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalStrictlyEquals call unimplemented")
+})
+
+var EmvalThrow = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalThrow call unimplemented")
+})
+
+var EmvalGreaterThan = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	// @todo: implement me.
+	panic("EmvalGreaterThan call unimplemented")
+})
