+++ internal/void.go
@@ -2,6 +2,7 @@ package embind
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/jerbob92/wazero-emscripten-embind/types"
 
@@ -28,3 +29,24 @@ func (vt *voidType) NativeType() api.ValueType {
 func (vt *voidType) GoType() string {
 	return ""
 }
+
+var RegisterVoid = api.GoModuleFunc(func(ctx context.Context, mod api.Module, stack []uint64) {
+	engine := MustGetEngineFromContext(ctx, mod).(*engine)
+
+	rawType := api.DecodeI32(stack[0])
+	name, err := engine.readCString(uint32(api.DecodeI32(stack[1])))
+	if err != nil {
+		panic(fmt.Errorf("could not read name: %w", err))
+	}
+
+	err = engine.registerType(rawType, &voidType{
+		baseType: baseType{
+			rawType:        rawType,
+			name:           name,
+			argPackAdvance: 0,
+		},
+	}, nil)
+	if err != nil {
+		panic(fmt.Errorf("could not register: %w", err))
+	}
+})
